Revert "[refactor] pinCount/PinCounts 객체 패키지 분리"
This reverts commit 038dcbc5
