+++ src/main/java/bowling/domain/State/FinalState.java
@@ -1,6 +1,5 @@
 package bowling.domain.State;
 
-import bowling.domain.PinCount.PinCount;
 import bowling.domain.score.FinishedScore;
 import bowling.domain.score.Score;
 import bowling.domain.score.UnDefinedScore;
