+++ src/main/java/bowling/domain/State/PinCount.java
@@ -1,4 +1,4 @@
-package bowling.domain.PinCount;
+package bowling.domain.State;
 
 import java.util.HashMap;
 import java.util.Map;
