+++ src/main/java/bowling/domain/State/Strike.java
@@ -1,6 +1,5 @@
 package bowling.domain.State;
 
-import bowling.domain.PinCount.PinCount;
 import bowling.domain.score.Score;
 import bowling.domain.score.UnFinishedScore;
