+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,7 +1,7 @@
 package bowling.domain.frame;
 
 import bowling.domain.State.FinalState;
-import bowling.domain.PinCount.PinCount;
+import bowling.domain.State.PinCount;
 import bowling.domain.State.Ready;
 import bowling.domain.State.State;
 import bowling.domain.score.Score;
